Carts for reg users

+++ backend/index.js
@@ -81,18 +81,35 @@ app.get('/api/registration', (req, res) => {;
 // 1.1 Add new user/Registration (Hashed password comrade, wake up Alan Turing to decrypt)
 
 app.post('/api/registration/add', async (req, res) => {
+    //create personal cart in DB
+    let stmtCreateCart, infoCreateCart;
+    try {
+        stmtCreateCart = dbWatches.prepare(`
+            INSERT INTO carts
+            DEFAULT VALUES
+        `);
+        infoCreateCart = stmtCreateCart.run();
+        if (infoCreateCart.lastInsertRowid <= 0) {
+            throw new Error();
+        }
+    } catch (error) {
+        res.status(409).json({message: 'Kunde inte skapa en personlig kundvagn'});
+    }
+    //later insert that cartId into registration-table as foreign key
+
     const saltRounds = 10;
     // Generate salt & hash password
     const hashed = await bcryptjs.hash(req.body.password, saltRounds);
 
    try{
     const stmt = dbWatches.prepare(`
-     INSERT INTO registration (user_name, password)
-     VALUES (:user_name,'${hashed}')
+     INSERT INTO registration (user_name, password, cartId)
+     VALUES (:user_name,'${hashed}', :cartId)
          `);
          const result = stmt.run({
              user_name: req.body.user_name,
-             password: req.body.password
+             password: req.body.password,
+             cartId: infoCreateCart.lastInsertRowid
             });
             res.json(result);
    } catch(err) {
@@ -111,11 +128,12 @@ app.get('/api/registration/:user_name/:password', async (req, res) => {
     const user = stmt.all({
         user_name: userNameDecoded,
     });
-    console.log(user[0]);
+    // console.log('Login response object:');
+    // console.log(user[0]);
     let validPassword = await bcryptjs.compare(passwordDecoded, user[0].password);
   
-    return res.json({ loginSuccess: validPassword, user_name: user[0].user_name });
-});
+    return res.json({ loginSuccess: validPassword, user_name: user[0].user_name, userCartId: user[0].cartId });
+    });
 
 // ......BCRYPT
 
@@ -152,7 +170,40 @@ const getUncryptedCartId = (cryptedId) => {
     return cartIdFromDB;
 }
 
-// Create a new cart
+const handleCartIdBeforeDB = (reqBody, reqParams) => {
+    let bodyOrParams;
+    for (let i in reqBody) {
+        if (i === 'cartId') {
+            bodyOrParams = 'body';
+        }
+    }
+    if (bodyOrParams !== 'body') {
+        for (let i in reqParams) {
+            if (i === 'cartid') {
+                bodyOrParams = 'params';
+            }
+        }
+    }
+
+    let cartIdFromFrontend;
+    if (bodyOrParams === 'body') {
+        cartIdFromFrontend = reqBody.cartId;
+    } else {
+        cartIdFromFrontend = decodeURIComponent(reqParams.cartid);
+    }
+
+    let plainCartId;
+    const regexOnlyNumbers = new RegExp('^[0-9]+$', 'gi');
+    if (regexOnlyNumbers.test(cartIdFromFrontend)) {
+        plainCartId = cartIdFromFrontend;
+    } else {
+        plainCartId = getUncryptedCartId(cartIdFromFrontend);
+    }
+
+    return plainCartId;
+}
+
+// Create a new logged out cart
 app.post('/api/cart/new', (req, res) => {
     let stmt = dbWatches.prepare(`
         INSERT INTO carts
@@ -185,7 +236,7 @@ app.post('/api/cart/new', (req, res) => {
 // Add a product to the cart - or - add more of the same product (quantity)
 app.post('/api/cart/add/productid/:productid', (req, res) => {
     // First get the uncrypted cartId that belongs to the users crypted one
-    const cartIdFromDB = getUncryptedCartId(req.body.cartId);
+    const plainCartId = handleCartIdBeforeDB(req.body, req.params);
 
     // Then use the uncrypted one to make changes in DB
     let stmt = dbWatches.prepare(`
@@ -193,7 +244,7 @@ app.post('/api/cart/add/productid/:productid', (req, res) => {
         VALUES (:cartIdParam, :productIdParam)
     `);
     let info = stmt.run({
-        cartIdParam: cartIdFromDB,
+        cartIdParam: plainCartId,
         productIdParam: req.params.productid
     });
     res.json({'Additions made': info.changes})
@@ -201,8 +252,7 @@ app.post('/api/cart/add/productid/:productid', (req, res) => {
 
 // Get all items in your cart
 app.get('/api/cart/:cartid', (req, res) => {
-    const cryptedCartId = decodeURIComponent(req.params.cartid);
-    const cartIdFromDB = getUncryptedCartId(cryptedCartId);
+    let plainCartId = handleCartIdBeforeDB(req.body, req.params);
 
     let stmt = dbWatches.prepare(`
         SELECT products.*,
@@ -212,15 +262,14 @@ app.get('/api/cart/:cartid', (req, res) => {
         GROUP BY productId
     `);
     let result = stmt.all({
-        cartIdParam: cartIdFromDB
+        cartIdParam: plainCartId
     })
     res.json(result);
 })
 
 // Delete a product from cart, no matter the quantity of that product.
 app.delete('/api/cart/:cartid/removefromcart/productid/:productid', (req, res) => {
-    const cryptedCartId = decodeURIComponent(req.params.cartid);
-    const cartIdFromDB = getUncryptedCartId(cryptedCartId);
+    let plainCartId = handleCartIdBeforeDB(req.body, req.params);
 
     let stmt = dbWatches.prepare(`
         DELETE FROM cartItems
@@ -228,7 +277,7 @@ app.delete('/api/cart/:cartid/removefromcart/productid/:productid', (req, res) =
         AND productId = :productIdParam
     `);
     let result = stmt.run({
-        cartIdParam: cartIdFromDB,
+        cartIdParam: plainCartId,
         productIdParam: req.params.productid
     })
     res.json(result);
@@ -236,8 +285,7 @@ app.delete('/api/cart/:cartid/removefromcart/productid/:productid', (req, res) =
 
 // Delete one, "Quantity - 1"
 app.delete('/api/cart/:cartid/quantitydecrease/productid/:productid', (req, res) => {
-    const cryptedCartId = decodeURIComponent(req.params.cartid);
-    const cartIdFromDB = getUncryptedCartId(cryptedCartId);
+    let plainCartId = handleCartIdBeforeDB(req.body, req.params);
 
     let stmt = dbWatches.prepare(`
         DELETE FROM cartItems
@@ -249,31 +297,57 @@ app.delete('/api/cart/:cartid/quantitydecrease/productid/:productid', (req, res)
         )
     `);
     let result = stmt.run({
-        cartIdParam: cartIdFromDB,
+        cartIdParam: plainCartId,
         productIdParam: req.params.productid
     })
     res.json(result);
 })
 
-// When purchase is done, delete cart and all products in cart from DB
-app.delete('/api/cart/:cartid/removecartcompletely', (req, res) => {
-    const cryptedCartId = decodeURIComponent(req.params.cartid);
-    const cartIdFromDB = getUncryptedCartId(cryptedCartId);
+// On login, move any items in logged-out cart to personal cart
+app.patch('/api/cart/movetopersonalcart', (req, res) => {
+    const oldCartId = getUncryptedCartId(req.body.cartIdLocalSt);
+    const newCartId = req.body.cartIdPersonal;
+
+    let stmt = dbWatches.prepare(`
+        UPDATE cartItems
+        SET cartId = :newCartId
+        WHERE cartId = :oldCartId
+    `);
+    let result = stmt.run({
+        oldCartId: oldCartId,
+        newCartId: newCartId
+    })
+
+    res.json(result);
+})
+
+// When purchase is done, for not logged in users: 
+// delete cart and all products in cart from DB
+app.delete('/api/cart/:cartid/cartcheckoutdb', (req, res) => {
+    let plainCartId = handleCartIdBeforeDB(req.body, req.params);
+    const cartIdParams = decodeURIComponent(req.params.cartid);
+    const regexOnlyNumbers = new RegExp('^[0-9]+$', 'gi');
 
     let stmt = dbWatches.prepare(`
         DELETE FROM cartItems
         WHERE cartId = :cartIdParam
     `);
     let result = stmt.run({
-        cartIdParam: cartIdFromDB
+        cartIdParam: plainCartId
     })
-    let stmt2 = dbWatches.prepare(`
-        DELETE FROM carts
-        WHERE cartId = :cartIdParam
-    `);
-    let result2 = stmt2.run({
-        cartIdParam: cartIdFromDB
-    })
+
+    let stmt2, result2;
+    if (!regexOnlyNumbers.test(cartIdParams)) {
+        // oinloggad varukorg, deleta även själva varukorgen
+        stmt2 = dbWatches.prepare(`
+            DELETE FROM carts
+            WHERE cartId = :cartIdParam
+        `);
+        result2 = stmt2.run({
+            cartIdParam: plainCartId
+        })
+    }
+    
     res.json({"cartItems": result, "carts": result2});
 })
 //=================
